Events (#4)
Add subscriptions for TCP Emitter server

Subscribe
Event emitted when a `TCP Emitter client` subscribes to a TCP Emitter event. The listeners of this event are invoked with the `TCP Emitter client` & `event
name`.

Note when a `TCP Emitter client` requests to subscribe to an event that it is already subscribed to, this event is not emitted.

Unsubscribe
Event emitted when a `TCP Emitter client` unsubscribes to a TCP Emitter event.
The listeners of this event are invoked with the `TCP Emitter client` & `event name`.

Note when a `TCP Emitter client` requests to unsubscribe from an event that it
is not subscribed to, this event is not emitted.

Broadcast
Event emitted when a `TCP Emitter client` broadcasts to a TCP Emitter event. The listeners of this event are invoked with the `TCP Emitter client`, `event name`
& `broadcast arguments`.

Closes #1